+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This can also be enabled programmatically with `warnings.simplefilter('default',
 - additional linter / static code analyser in GitHub Actions CI pipeline: [semgrep](https://github.com/returntocorp/semgrep)
 ### Fixed
 - hyperlinks were not working on encrypted files - thanks to @andersonhc
+- a bug was introduced in the previous release (2.6.1): `FPDF.set_link()` could not update links generated with `add_link()`
 - unicode (non limited to ASCII) text can now be provided as metadata [#685](https://github.com/PyFPDF/fpdf2/issues/685)
 - all `TitleStyle` constructor parameters are now effectively optional
 - memory usage was reduced by 10 MiB in some cases, thanks to a small optimization in using `fonttools`